feat: Add read and delete methods for files

+++ src/main/java/com/tistory/project_api/repository/FilesRepository.java
@@ -3,5 +3,5 @@ package com.tistory.project_api.repository;
 import com.tistory.project_api.domain.entity.FilesEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilesRepository extends JpaRepository<FilesEntity, Integer> {
+public interface FilesRepository extends JpaRepository<FilesEntity, Long> {
 }
